move ch24 to ch29 to make room for etl chapters
