+++ src/ch29_person_viewer/_ref/ch29_ref.json
@@ -1,7 +1,7 @@
 {
   "chapter_blurb": "Tools for Visualizing PersonUnits", 
   "chapter_content": "Tools for Visualizing PersonUnits", 
-  "chapter_description": "ch24_person_viewer", 
-  "chapter_number": 24, 
+  "chapter_description": "ch29_person_viewer", 
+  "chapter_number": 29, 
   "ontology summary": "PersonUnits are complicated, graphic tools are useful."
 }
