+++ src/ch30_etl_app/test/test_z_app_runs.py
@@ -3,11 +3,7 @@ from os import environ as os_environ
 from pytest import mark as pytest_mark
 
 
-@pytest_mark.skipif(
-    not os_environ.get("DISPLAY"),
-    reason="No display available for Tkinter",
-)
-def test_ETLApp_ApplicationInitRunsWithoutError():
+def test_ETLApp_ApplicationInitRunsWithoutError(tk_app):
     # ESTABLISH / WHEN
     root = ETLApp()
     # Schedule close almost immediately
