+++ src/linter/chapter_move_main.py
@@ -1,6 +1,6 @@
 from ch00_py.file_toolbox import create_path, open_json, save_json
 from ch98_docs_builder._ref.ch98_path import create_chapter_ref_path
-from linter.chapter_migration_tools import (
+from linter.chapter_move_tools import (
     delete_if_empty_or_pycache_only,
     first_level_dirs_with_prefix,
     rename_files_and_dirs_4times,
@@ -29,7 +29,7 @@ from subprocess import (
 
 # HOW TO USE:
 # Open up CMD, change directory to repo
-# Enter this: python -m src.linter.ch_move
+# Enter this: python -m src.linter.chapter_move_main
 
 
 def first_level_dirs_with_prefix(path_prefix: str):
