+++ src/ref/keg_definitions.json
@@ -77,7 +77,7 @@
   "ch21": "Defines WorldDirs that admin MomentUnits input data and outcomes", 
   "ch22": "Defines Lobby Tools for gifts, World Scenarios.", 
   "ch23": "Defines Finance Tools for ", 
-  "ch24": "Tools for Visualizing PersonUnits", 
+  "ch29": "Tools for Visualizing PersonUnits", 
   "ch30": "ETL app to create day punches from excel files.", 
   "ch98": "Defines Tools that create documentation.", 
   "Chiday": "Third Day of 5 day week of 'Five' Epoch",
