+++ src/ref/keg_exam.json
@@ -7,7 +7,7 @@
 "5": {"question_type": "Keyword Definition", "question_str": "Have you read the Kegology definition of 'totality'?", "keyword": "totality"},
 "6": {"question_type": "Keyword Definition", "question_str": "Have you read the Kegology definition of 'ch98'?", "keyword": "ch98"},
 "7": {"question_type": "Keyword Definition", "question_str": "Have you read the Kegology definition of 'ch30'?", "keyword": "ch30"},
-"8": {"question_type": "Keyword Definition", "question_str": "Have you read the Kegology definition of 'ch24'?", "keyword": "ch24"},
+"8": {"question_type": "Keyword Definition", "question_str": "Have you read the Kegology definition of 'ch29'?", "keyword": "ch29"},
 "9": {"question_type": "Keyword Definition", "question_str": "Have you read the Kegology definition of 'ch23'?", "keyword": "ch23"},
 "10": {"question_type": "Keyword Definition", "question_str": "Have you read the Kegology definition of 'finance'?", "keyword": "finance"},
 "11": {"question_type": "Keyword Definition", "question_str": "Have you read the Kegology definition of 'financial'?", "keyword": "financial"},
