+++ src/ref/keywords_main.json
@@ -113,7 +113,7 @@
   "ch21": {"exam_tier": 0, "init_chapter": "ch21"}, 
   "ch22": {"exam_tier": 0, "init_chapter": "ch22"}, 
   "ch23": {"exam_tier": 0, "init_chapter": "ch23"}, 
-  "ch24": {"exam_tier": 0, "init_chapter": "ch24"}, 
+  "ch29": {"exam_tier": 0, "init_chapter": "ch29"}, 
   "ch30": {"exam_tier": 0, "init_chapter": "ch30"}, 
   "ch98": {"exam_tier": 0, "init_chapter": "ch98"}, 
   "class_type": {"exam_tier": 0, "init_chapter": "ch07"},
